packages/qlego-mqt-verification/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qlego-mqt-verification"
version = "0.1.0"
dependencies = [
    "qlego-core",
    "mqt.qcec"
]

[tool.setuptools.packages.find]
where = ["src"]
